Move main screen navigation logic to MainActivity
