+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/keystore/KeyStoreActivity.kt
@@ -10,7 +10,14 @@ import androidx.biometric.BiometricManager.Authenticators.DEVICE_CREDENTIAL
 import androidx.biometric.BiometricPrompt
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
@@ -26,7 +33,7 @@ import androidx.compose.ui.window.Dialog
 import androidx.core.content.ContextCompat
 import io.horizontalsystems.bankwallet.R
 import io.horizontalsystems.bankwallet.core.BaseActivity
-import io.horizontalsystems.bankwallet.modules.launcher.LaunchModule
+import io.horizontalsystems.bankwallet.modules.main.MainModule
 import io.horizontalsystems.bankwallet.ui.compose.ComposeAppTheme
 import io.horizontalsystems.bankwallet.ui.compose.components.BottomSheetsElementsButtons
 import io.horizontalsystems.bankwallet.ui.compose.components.BottomSheetsElementsHeader
@@ -127,7 +134,7 @@ private fun KeyStoreScreen(
 ) {
     if (viewModel.openMainModule) {
         viewModel.openMainModuleCalled()
-        LaunchModule.start(LocalContext.current)
+        MainModule.startAsNewTask(LocalContext.current)
     }
 
     if (viewModel.closeApp) {
