--- app/src/main/java/io/horizontalsystems/bankwallet/modules/launcher/LaunchModule.kt
@@ -1,24 +1,0 @@
-package io.horizontalsystems.bankwallet.modules.launcher
-
-import android.content.Context
-import android.content.Intent
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.ViewModelProvider
-import io.horizontalsystems.bankwallet.core.App
-
-object LaunchModule {
-
-    class Factory : ViewModelProvider.Factory {
-        @Suppress("UNCHECKED_CAST")
-        override fun <T : ViewModel> create(modelClass: Class<T>): T {
-            return LaunchViewModel(App.accountManager, App.pinComponent, App.systemInfoManager, App.keyStoreManager, App.localStorage) as T
-        }
-    }
-
-    fun start(context: Context) {
-        val intent = Intent(context, LauncherActivity::class.java)
-        intent.flags = Intent.FLAG_ACTIVITY_NEW_TASK or Intent.FLAG_ACTIVITY_CLEAR_TASK
-        context.startActivity(intent)
-    }
-
-}
