+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/main/MainModule.kt
@@ -40,6 +40,12 @@ object MainModule {
         context.startActivity(intent)
     }
 
+    fun startAsNewTask(context: Context) {
+        val intent = Intent(context, MainActivity::class.java)
+        intent.flags = Intent.FLAG_ACTIVITY_NEW_TASK or Intent.FLAG_ACTIVITY_CLEAR_TASK
+        context.startActivity(intent)
+    }
+
     fun startAsNewTask(context: Activity) {
         val intent = Intent(context, MainActivity::class.java)
         intent.flags = Intent.FLAG_ACTIVITY_NEW_TASK or Intent.FLAG_ACTIVITY_CLEAR_TASK
